// src/Agent/Agent.Portal/Client/src/Common/Contracts/Permissions.ts
export enum RBACRoleNames {
    contributor = 'Contributor',
    containerAppsContributor = 'ContainerAppsContributor',
    logAnalyticsReader = 'LogAnalyticsReader',
    websitesContributor = 'WebsitesContributor',
    webPlanContributor = 'WebPlanContributor',
    reader = 'Reader',
    containerAppsOperator = 'ContainerAppsOperator',
    azureKubernetesServiceRbacReader = 'AzureKubernetesServiceRbacReader',
    azureKubernetesServiceClusterUser = 'AzureKubernetesServiceClusterUser',
    azureKubernetesServiceClusterAdmin = 'AzureKubernetesServiceClusterAdmin',
    azureKubernetesServiceRbacClusterAdmin = 'AzureKubernetesServiceRbacClusterAdmin',
    azureMonitorMonitoringContributor = 'AzureMonitorMonitoringContributor',
    applicationInsightsComponentContributor = 'ApplicationInsightsComponentContributor',
    logAnalyticsContributor = 'LogAnalyticsContributor',
    storageBlobDataContributor = 'StorageBlobDataContributor',
    documentDbAccountContributor = 'DocumentDbAccountContributor',
    redisCacheContributor = 'RedisCacheContributor',
    sqlDbContributor = 'SqlDbContributor',
    monitoringReader = 'MonitoringReader',
    storageBlobDataReader = 'StorageBlobDataReader',
    virtualMachineContributor = 'VirtualMachineContributor',
    azureDatabaseForPostgreSqlContributor = 'AzureDatabaseForPostgreSqlContributor',
    sqlServerContributor = 'SqlServerContributor',
    storageAccountContributor = 'StorageAccountContributor',
    postgreSqlFlexibleServerLongTermRetentionBackupRole = 'PostgreSQLFlexibleServerLongTermRetentionBackupRole',
    sqlManagedInstanceContributor = 'SqlManagedInstanceContributor',
    dataFactoryContributor = 'DataFactoryContributor',
    hdInsightOnAksClusterAdmin = 'HDInsightOnAKSClusterAdmin',
    hdInsightOnAksClusterPoolAdmin = 'HDInsightOnAKSClusterPoolAdmin',
    azureMlComputeOperator = 'AzureMLComputeOperator',
    azureMlDataScientist = 'AzureMLDataScientist',
    cognitiveServicesContributor = 'CognitiveServicesContributor',
    cognitiveServicesOpenAiContributor = 'CognitiveServicesOpenAIContributor',
    cognitiveServicesCustomVisionContributor = 'CognitiveServicesCustomVisionContributor',
    cognitiveServicesLanguageWriter = 'CognitiveServicesLanguageWriter',
    cognitiveServicesLuisWriter = 'CognitiveServicesLUISWriter',
    cognitiveServicesQnaMakerEditor = 'CognitiveServicesQnAMakerEditor',
    cognitiveServicesSpeechContributor = 'CognitiveServicesSpeechContributor',
    healthcareAgentEditor = 'HealthcareAgentEditor',
    searchServiceContributor = 'SearchServiceContributor',
    azureDigitalTwinsDataOwner = 'AzureDigitalTwinsDataOwner',
    deviceProvisioningServiceDataContributor = 'DeviceProvisioningServiceDataContributor',
    deviceUpdateAdministrator = 'DeviceUpdateAdministrator',
    iotHubDataContributor = 'IoTHubDataContributor',
    iotHubRegistryContributor = 'IoTHubRegistryContributor',
    iotHubTwinContributor = 'IoTHubTwinContributor',
    apiManagementServiceContributor = 'APIManagementServiceContributor',
    apiManagementServiceOperatorRole = 'APIManagementServiceOperatorRole',
    apiManagementWorkspaceContributor = 'APIManagementWorkspaceContributor',
    appConfigurationContributor = 'AppConfigurationContributor',
    azureServiceBusDataOwner = 'AzureServiceBusDataOwner',
    logicAppContributor = 'LogicAppContributor',
    monitoringContributor = 'MonitoringContributor',
    workbookContributor = 'WorkbookContributor',
    azureCenterForSapSolutionsAdministrator = 'AzureCenterForSAPSolutionsAdministrator',
    costManagementContributor = 'CostManagementContributor',
    hdInsightClusterOperator = 'HDInsightClusterOperator',
    cognitiveServicesCustomVisionReader = 'CognitiveServicesCustomVisionReader',
    cognitiveServicesDataReader = 'CognitiveServicesDataReader',
    cognitiveServicesLanguageReader = 'CognitiveServicesLanguageReader',
    cognitiveServicesLuisReader = 'CognitiveServicesLUISReader',
    cognitiveServicesQnaMakerReader = 'CognitiveServicesQnAMakerReader',
    cognitiveServicesUsagesReader = 'CognitiveServicesUsagesReader',
    searchIndexDataReader = 'SearchIndexDataReader',
    azureDigitalTwinsDataReader = 'AzureDigitalTwinsDataReader',
    deviceProvisioningServiceDataReader = 'DeviceProvisioningServiceDataReader',
    deviceUpdateReader = 'DeviceUpdateReader',
    iotHubDataReader = 'IoTHubDataReader',
    apiManagementServiceReader = 'APIManagementServiceReader',
    apiManagementWorkspaceReader = 'APIManagementWorkspaceReader',
    appConfigurationReader = 'AppConfigurationReader',
    logicAppOperator = 'LogicAppOperator',
    workbookReader = 'WorkbookReader',
    azureCenterForSapSolutionsReader = 'AzureCenterForSAPSolutionsReader',
    costManagementReader = 'CostManagementReader',
}